% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_update.R
\name{add_collabs}
\alias{add_collabs}
\title{Adding collaborators to a repository}
\usage{
add_collabs(token = NULL, owner, repo, user, permission)
}
\arguments{
\item{token}{a github pat}

\item{owner}{owner of the repository}

\item{repo}{name of the repository}

\item{user}{user being added to repository}

\item{permission}{can be one of: pull, triage, push, maintain, admin}
}
\description{
Adding collaborators to a repository
}
\examples{
\dontrun{add_collabs(owner = GITHUBREPOOWNER, repo = REPONAME,
user = GITHUBUSERNAME, permission = "admin")}
}
