% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_gh_login}
\alias{get_gh_login}
\title{Get github login}
\usage{
get_gh_login(gh_pat = NULL)
}
\arguments{
\item{gh_pat}{character, a github token
default: GITHUB_PAT}
}
\description{
Get github login
}
