% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_list.R
\name{list_collabs}
\alias{list_collabs}
\title{List usernames of collaborators for repository}
\usage{
list_collabs(token = NULL, owner, repo, affiliation)
}
\arguments{
\item{token}{a github pat}

\item{owner}{github username}

\item{repo}{github repository}

\item{affiliation}{can be one of: outside, direct, all}
}
\value{
collaborators of a repository
}
\description{
List usernames of collaborators for repository
}
\examples{
\dontrun{list_collabs(owner = GITHUBREPOOWNER, repo = REPONAME, affiliation = "all")}
}
