% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_list.R
\name{list_repositories}
\alias{list_repositories}
\title{List all repositories which the authenticated user has access to}
\usage{
list_repositories(token = NULL, visibility = "all")
}
\arguments{
\item{token}{a github personal access token on an env var storing it.
If `NULL` (the default) the function tries to fetch the token from `GITHUB_PAT`
env var.}

\item{visibility}{can be one of: private, public, or all}
}
\value{
a data frame of repositories
}
\description{
List all repositories which the authenticated user has access to
}
\examples{
\dontrun{list_repositories(visibility = "all")}
}
