% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_update.R
\name{update_desc}
\alias{update_desc}
\title{Updating Description of some repository}
\usage{
update_desc(token = NULL, user, repo, desc)
}
\arguments{
\item{token}{a github pat}

\item{user}{github username}

\item{repo}{github repository}

\item{desc}{the updated description}
}
\description{
Updating Description of some repository
}
\examples{
\dontrun{update_desc(user = GITHUBUSERNAME, repo = REPONAME, desc = "new description")}
}
