% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wrangle_repo_data}
\alias{wrangle_repo_data}
\title{Wrangle repository data
Helper function to return a tidy data frame of
repository data.}
\usage{
wrangle_repo_data(gh_response)
}
\arguments{
\item{gh_response}{a response from github}
}
\description{
Wrangle repository data
Helper function to return a tidy data frame of
repository data.
}
